Major refactor of the schedulers, builders, and master.cfg
This was done because while the code in this repo looked acceptable, the generated
code was completely illegible, which made debugging remarkably difficult.  This changeset
causes less to be hardcoded, and more to be built at runtime.  It's also more maintainable
since the majority of the Ansible interpolation shows up in master.cfg, rather than scattered
around in the builders and schedulers as well.