Fix NGINX config for CORS preflight requests
CORS preflight requests are only accepted by browsers if the server
responds with a 2xx code. Some browsers even only accept 200 as response
response code, meaning we can't use 204. Before this commit, the request
was handled by Opencast. This usually lead to a redirect because most
endpoints are protected and Opencast wanted to redirect to the login page.