Continued work on parallel building.  Further message below.
This commit does a lot of the backend work required to properly support multiple builds
in parallel on the same worker.  The issue is that the job dispatching logic does not
understand that a given worker might be debian, or might be centos based, at any given time.

If you enable parallel builds (see master.cfg) you will end up with builds being scheduled to
a worker regardless of its underlying distro - so debian package building on a redhat box, or
vice versa.  We need to make the dispatching smarter before we can enable this.